pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Spotify-Translator"
version = "0.0.1"
authors = [
  { name="Hajin Park", email="hajin1819@gmail.com" },
]
description = "Generate lyric translations and transcriptions from Spotify URLs using OpenAI's Whisper model."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "openai-whisper",
  "spotify-dl",
  "spotipy"
]
license = {file = "LICENSE"}
keywords = ["spotify", "translate", "translator", "translation", "transcript", "transcription", "transcriptor", "youtube", "yt", "OpenAI", "Whisper"]

[project.urls]
Homepage = "https://github.com/hajin-park/spotify-translator"
Issues = "https://github.com/hajin-park/spotify-translator/issues"